moved to internal; no underscore in the name; changelog entry
