+++ internal/ingestionmethod/ingestion_method.go
@@ -2,7 +2,7 @@
 // or more contributor license agreements. Licensed under the Elastic License;
 // you may not use this file except in compliance with the Elastic License.
 
-package ingestion_method
+package ingestionmethod
 
 import (
 	"fmt"
