+++ packages/package.go
@@ -19,7 +19,7 @@ import (
 	"github.com/elastic/go-ucfg/yaml"
 
 	"github.com/elastic/package-registry/categories"
-	"github.com/elastic/package-registry/ingestion_method"
+	"github.com/elastic/package-registry/internal/ingestionmethod"
 )
 
 const (
@@ -30,7 +30,7 @@ const (
 
 var (
 	Categories       = categories.DefaultCategories()
-	IngestionMethods = ingestion_method.DefaultIngestionMethod()
+	IngestionMethods = ingestionmethod.DefaultIngestionMethod()
 
 	// Deprecated, keeping for backwards compatibility, Categories should be used instead.
 	CategoryTiles = categoryTitles(categories.DefaultCategories())
